Working on Homw screen

+++ app/src/main/java/grand/app/moon/presentation/home/models/ItemHomeRV.kt
@@ -47,7 +47,7 @@ data class ItemAdvertisementInResponseHome(
 	var country: Country?, // name
 	var city: City?, // name
 	var store: Store?, // stories, id, image, nicname
-	var price: Int?,
+	var price: Float?,
 	@SerializedName("is_negotiable") var negotiable: Int?,
 ) : ItemInResponseHome {
